requirements.txt
joblib==1.2.0
matplotlib==3.6.3
numpy==1.24.2
pandas==1.5.3
python_dateutil==2.8.2
scikit_learn==1.2.2
seaborn==0.12.2
